Generate a .Rmd file containing code to perform differential expression analysis with phylolimma

A function to generate code that can be run to perform differential expression analysis of RNAseq data (comparing two conditions) by applying a length normalisation and transformation followed by differential expression analysis with phylolimma. The code is written to a .Rmd file. This function is generally not called by the user, the main interface for performing differential expression analysis is the runDiffExp function.

phylolimma.createRmd(
  data.path,
  result.path,
  codefile,
  norm.method,
  model = "BM",
  measurement_error = TRUE,
  extraDesignFactors = NULL,
  lengthNormalization = "RPKM",
  dataTransformation = "log2",
  ...
)

Arguments

data.path
The path to a .rds file containing the compData object that will be used for the differential expression analysis.

result.path
The path to the file where the result object will be saved.

codefile
The path to the file where the code will be written.

norm.method
The between-sample normalization method used to compensate for varying library sizes and composition in the differential expression analysis. The normalization factors are calculated using the calcNormFactors of the edgeR package. Possible values are "TMM", "RLE", "upperquartile" and "none"

model
The model for trait evolution on the tree. Default to "BM".

measurement_error
A logical value indicating whether there is measurement error. Default to TRUE.

extraDesignFactors
A vector containing the extra factors to be passed to the design matrix of limma. All the factors need to be a sample.annotations from the compData object. It should not contain the "condition" factor column, that will be added automatically.

lengthNormalization
one of "none" (no correction), "TPM" or "RPKM" (default).

dataTransformation
one of "log2", "asin(sqrt)" or "sqrt". Data transformation to apply to the normalized data.

...
Further arguments to be passed to function phylolimma.

Value

The function generates a .Rmd file containing the code for performing the differential expression analysis. This file can be executed using e.g. the knitr package.

Details

For more information about the methods and the interpretation of the parameters, see the phylolimma package and the corresponding publications.

The length.matrix field of the compData object is used to normalize the counts, using the normalisation specified by lengthNormalization.

References

Smyth GK (2005): Limma: linear models for microarray data. In: 'Bioinformatics and Computational Biology Solutions using R and Bioconductor'. R. Gentleman, V. Carey, S. Dudoit, R. Irizarry, W. Huber (eds), Springer, New York, pages 397-420

Musser, JM, Wagner, GP. (2015): Character trees from transcriptome data: Origin and individuation of morphological characters and the so‐called “species signal”. J. Exp. Zool. (Mol. Dev. Evol.) 324B: 588– 604.

Author

Charlotte Soneson

Examples

try(
if (require(phylolimma)) {
tmpdir <- normalizePath(tempdir(), winslash = "/")
mydata.obj <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                    samples.per.cond = 5, n.diffexp = 100, 
                                    id.species = factor(1:10),
                                    lengths.relmeans = rpois(1000, 1000),
                                    lengths.dispersions = rgamma(1000, 1, 1),
                                    output.file = file.path(tmpdir, "mydata.rds"))
runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "length.limma", 
           Rmdfunction = "phylolimma.createRmd", 
           output.directory = tmpdir, norm.method = "TMM")
})

#> Loading required package: phylolimma

#> Warning: there is no package called ‘phylolimma’