Generate a .Rmd file containing code to perform differential expression analysis with limma after the variance-stabilizing transformation provided in DESeq2

A function to generate code that can be run to perform differential expression analysis of RNAseq (comparing two conditions) by applying the variance-stabilizing transformation of the DESeq2 package followed by differential expression analysis using limma. The code is written to a .Rmd file. This function is generally not called by the user, the main interface for performing differential expression analysis is the runDiffExp function.

vst.length.limma.createRmd(
  data.path,
  result.path,
  codefile,
  fit.type,
  extraDesignFactors = NULL
)

Arguments

data.path
The path to a .rds file containing the compData object that will be used for the differential expression analysis.

result.path
The path to the file where the result object will be saved.

codefile
The path to the file where the code will be written.

fit.type
The fitting method used to get the dispersion-mean relationship. Possible values are "parametric" and "local".

extraDesignFactors
A vector containing the extra factors to be passed to the design matrix of DESeq2. All the factors need to be a sample.annotations from the compData object. It should not contain the "condition" factor column, that will be added automatically.

Value

The function generates a .Rmd file containing the code for performing the differential expression analysis. This file can be executed using e.g. the knitr package.

Details

For more information about the methods and the interpretation of the parameters, see the limma and DESeq2 packages and the corresponding publications.

References

Smyth GK (2005): Limma: linear models for microarray data. In: 'Bioinformatics and Computational Biology Solutions using R and Bioconductor'. R. Gentleman, V. Carey, S. Dudoit, R. Irizarry, W. Huber (eds), Springer, New York, pages 397-420

Anders S and Huber W (2010): Differential expression analysis for sequence count data. Genome Biology 11:R106

Author

Charlotte Soneson, Paul Bastide, Mélina Gallopin

Examples

try(
if (require(limma)) {
tmpdir <- normalizePath(tempdir(), winslash = "/")
mydata.obj <- generateSyntheticData(dataset = "mydata", n.vars = 1000, 
                                    samples.per.cond = 5, n.diffexp = 100, 
                                    output.file = file.path(tmpdir, "mydata.rds"))
runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "vst.length.limma", 
           Rmdfunction = "vst.length.limma.createRmd", 
           output.directory = tmpdir, fit.type = "parametric")
})

#> 
#> 
#> processing file: /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpfwx6vA/tempcode754f5b3676b5.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |...................................                                   |  50%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................................................................| 100%
#> label: unnamed-chunk-1 (with options) 
#> List of 6
#>  $ echo   : logi TRUE
#>  $ eval   : logi TRUE
#>  $ include: logi TRUE
#>  $ message: logi FALSE
#>  $ error  : logi TRUE
#>  $ warning: logi TRUE
#> 

#> converting counts to integer mode

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> 

#> output file: /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpfwx6vA/tempcode754f5b3676b5.md

#> [1] TRUE